Proof of concept for TilingRenderer.
SimpleDisplay is a small (probably temporary) helper class to test things.
Invoked with the arguments "p6m 7 3 2" it demonstrates the general
feasibility of the TilingRenderer in terms of speed and correctness. For
numerical reasons the displayed image has some seams, but that is to be
expected.

P6m is the first group to be implemented, based on the simple them of a
triangle reflection group. Others will follow, many of them based on kites.
For these, there are alreay some classes in place, but untested so far, and
quite probably buggy.